chore: move files to the internal folder
